Rename to `invite-manager.ts`.
